// src/generated/models/Svm_pt_owneroccupanciesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Svm_pt_owneroccupanciesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Svm_pt_owneroccupanciesstatecode = keyof typeof Svm_pt_owneroccupanciesstatecode;
export const Svm_pt_owneroccupanciesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Svm_pt_owneroccupanciesstatuscode = keyof typeof Svm_pt_owneroccupanciesstatuscode;

export interface Svm_pt_owneroccupanciesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Svm_pt_owneroccupanciesstatecode;
  statuscode?: Svm_pt_owneroccupanciesstatuscode;
  svm_pt_adults?: number;
  svm_pt_babies?: number;
  svm_pt_children?: number;
  svm_pt_fromdate?: string;
  svm_pt_name?: string;
  svm_pt_owneroccupancyid: string;
  "svm_pt_Property@odata.bind"?: string;
  svm_pt_todate?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Svm_pt_owneroccupancies extends Svm_pt_owneroccupanciesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  svm_pt_propertyname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  svm_pt_property?: object;
  _svm_pt_property_value?: string;
}